requirements.txt
appnope==0.1.3
asttokens==2.1.0
attrs==22.1.0
autopep8==2.0.2
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
coloredlogs==15.0.1
commonmark==0.9.1
debugpy==1.6.3
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.5
docutils==0.19
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.7.1
Flask==2.2.2
freetype-py==2.1.0
html2text==2020.1.16
humanfriendly==10.0
idna==3.3
imageio==2.19.5
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
ipykernel==6.17.1
ipython==8.6.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
keyring==23.11.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
modbus-tk==1.1.2
more-itertools==9.0.0
moviepy==1.0.3
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.1
nest-asyncio==1.5.6
numpy==1.19.3
opencv-python==4.6.0.66
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pinpong==0.4.9
pipenv==2022.7.4
pkginfo==1.8.3
platformdirs==2.5.2
proglog==0.1.10
progress==1.6
prompt-toolkit==3.0.32
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
pyzmq==24.0.1
readme-renderer==37.3
requests==2.28.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==12.6.0
six==1.16.0
some-package==0.1
soupsieve==2.3.2.post1
stack-data==0.6.1
teedoc==1.32.0
tinycss2==1.1.1
tomli==2.0.1
torch==1.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.4.0
twine==4.0.1
typing_extensions==4.3.0
urllib3==1.26.10
virtualenv==20.15.1
virtualenv-clone==0.5.7
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
zipp==3.8.1
